add tests for 'has_session_expired()'

+++ rasa/core/events/__init__.py
@@ -209,7 +209,7 @@ class UserUttered(Event):
         intent=None,
         entities=None,
         parse_data: Optional[Dict[Text, Any]] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         input_channel: Optional[Text] = None,
         message_id: Optional[Text] = None,
         metadata: Optional[Dict] = None,
@@ -237,7 +237,7 @@ class UserUttered(Event):
     def _from_parse_data(
         text: Text,
         parse_data: Dict[Text, Any],
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         input_channel: Optional[Text] = None,
         message_id: Optional[Text] = None,
         metadata: Optional[Dict] = None,
@@ -441,7 +441,7 @@ class SlotSet(Event):
         self,
         key: Text,
         value: Optional[Any] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.key = key
@@ -596,7 +596,7 @@ class ReminderScheduled(Event):
         trigger_date_time: datetime,
         name: Optional[Text] = None,
         kill_on_user_message: bool = True,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ):
         """Creates the reminder
@@ -684,7 +684,7 @@ class ReminderCancelled(Event):
     def __init__(
         self,
         action_name: Text,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ):
         """
@@ -758,7 +758,7 @@ class StoryExported(Event):
     def __init__(
         self,
         path: Optional[Text] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ):
         self.path = path
@@ -800,7 +800,7 @@ class FollowupAction(Event):
     def __init__(
         self,
         name: Text,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.action_name = name
@@ -906,7 +906,7 @@ class ActionExecuted(Event):
         action_name: Text,
         policy: Optional[Text] = None,
         confidence: Optional[float] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict] = None,
     ):
         self.action_name = action_name
@@ -975,7 +975,7 @@ class AgentUttered(Event):
         self,
         text: Optional[Text] = None,
         data: Optional[Any] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.text = text
@@ -1038,7 +1038,7 @@ class Form(Event):
     def __init__(
         self,
         name: Optional[Text],
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.name = name
@@ -1089,7 +1089,7 @@ class FormValidation(Event):
     def __init__(
         self,
         validate: bool,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.validate = validate
@@ -1134,7 +1134,7 @@ class ActionExecutionRejected(Event):
         action_name: Text,
         policy: Optional[Text] = None,
         confidence: Optional[float] = None,
-        timestamp: Optional[int] = None,
+        timestamp: Optional[float] = None,
         metadata: Optional[Dict[Text, Any]] = None,
     ) -> None:
         self.action_name = action_name
